+++ .github/dependabot.yml
@@ -1,0 +1,30 @@
+version: 2
+updates:
+  - package-ecosystem: "maven"
+    directory: "/"
+    schedule:
+      interval: "weekly"
+      day: "monday"
+    open-pull-requests-limit: 10
+    groups:
+      maven-plugins:
+        patterns:
+          - "org.apache.maven.plugins:*"
+      test-dependencies:
+        patterns:
+          - "org.junit*"
+          - "junit:junit"
+      jacoco:
+        patterns:
+          - "org.jacoco:*"
+    ignore:
+      # Ignore patch updates for Maven plugins to reduce noise
+      - dependency-name: "org.apache.maven.plugins:*"
+        update-types: ["version-update:semver-patch"]
+
+  - package-ecosystem: "github-actions"
+    directory: "/"
+    schedule:
+      interval: "weekly"
+      day: "monday"
+    open-pull-requests-limit: 10
