+++ jacoco-console-reporter/pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>com.github</groupId>
+        <groupId>io.github</groupId>
         <artifactId>jacoco-console-reporter-parent</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
