+++ jacoco-console-reporter/src/test/java/io/github/svaningelgem/JacocoConsoleReporterMojoTest.java
@@ -1,9 +1,7 @@
-package com.github;
+package io.github.svaningelgem;
 
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.testing.MojoRule;
-import org.jetbrains.annotations.Contract;
-import org.jetbrains.annotations.NotNull;
 import org.junit.Rule;
 import org.junit.Test;
